202502010_문의화면 모바일 기능 수정(답글달린글 삭제 불가)

+++ src/main/java/mes/app/MobileUsr/AppInfo/QnA/QnAController.java
@@ -60,7 +60,14 @@ public class QnAController {
             Optional<TB_USERQST> answer = qnAService.AnswerSearchBy(tbUserqst.getQSTSEQ());
 
             List<TB_USERQST> items = new ArrayList<>();
-            answer.ifPresent(items::add);
+
+            if(answer.isPresent()){
+
+                result.success = false;
+                result.message = "답변이 달린 글은 삭제가 불가능 합니다.";
+                return result;
+            }
+
             items.add(tbUserqst);
 
             qnAService.deleteQna(items);
